pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "habittracker"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "prompt_toolkit==3.0.52",
  "python-dateutil==2.9.0.post0",
  "requests==2.32.5"
]

[project.scripts]
habittracker = "habittracker.main:main"

[project.optional-dependencies]
dev = [
  "black==25.12.0",
  "flake8==7.3.0",
  "isort==7.0.0",
  "pytest==9.0.2"
]

[tool.setuptools.packages.find]
where = ["src"]
